Lab3 file added
Circles are created using Midpoint Circle Algo. 
On right clicking on the mouse, new circles are created with the coordinates of the clicked screen .
Circles continue to grow in size.
Growth rate increased by right key click and decreased by left key click.
On collision with the screen, circle disappear.
On clicking on the space bar, the entire process stops.